重构log

+++ doublekit-pipeline-server/src/main/java/com/doublekit/pipeline/instance/service/PipelineExecLogServiceImpl.java
@@ -4,13 +4,13 @@ import com.doublekit.beans.BeanMapper;
 import com.doublekit.join.JoinTemplate;
 import com.doublekit.pipeline.instance.dao.PipelineExecLogDao;
 import com.doublekit.pipeline.instance.entity.PipelineExecLogEntity;
-import com.doublekit.pipeline.instance.model.PipelineExecHistory;
-import com.doublekit.pipeline.instance.model.PipelineExecLog;
+import com.doublekit.pipeline.instance.model.*;
 import com.doublekit.rpc.annotation.Exporter;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.Map;
 
 @Service
 @Exporter
@@ -32,26 +32,56 @@ public class PipelineExecLogServiceImpl implements PipelineExecLogService {
         return pipelineExecLogDao.createLog(pipelineExecLogEntity);
     }
 
+    //创建
+    @Override
+    public PipelineExecLog createLog(){
+        PipelineExecLog pipelineExecLog = new PipelineExecLog();
+        Map<String, String> map = pipelineCodeLogService.createCodeLog();
+
+        PipelineCodeLog pipelineCodeLog = new PipelineCodeLog();
+        pipelineCodeLog.setLogCodeId(map.get("codeLogId"));
+        pipelineExecLog.setCodeLog(pipelineCodeLog);
+
+        PipelineTestLog pipelineTestLog = new PipelineTestLog();
+        pipelineTestLog.setLogTestId(map.get("testLogId"));
+        pipelineExecLog.setTestLog(pipelineTestLog);
+
+        PipelineStructureLog pipelineStructureLog = new PipelineStructureLog();
+        pipelineStructureLog.setLogStructureId(map.get("structureLogId"));
+        pipelineExecLog.setStructureLog(pipelineStructureLog);
+
+        PipelineDeployLog pipelineDeployLog = new PipelineDeployLog();
+        pipelineDeployLog.setLogDeployId(map.get("deployLogId"));
+        pipelineExecLog.setDeployLog(pipelineDeployLog);
+
+        String logId = createLog(pipelineExecLog);
+        pipelineExecLog.setLogId(logId);
+        return pipelineExecLog;
+    }
+
+
     //删除
     @Override
-    public void deleteLog(String pipelineLogId) {
-        PipelineExecLog oneLog = findOneLog(pipelineLogId);
-        pipelineCodeLogService.deleteCodeLog(oneLog);
-        pipelineExecLogDao.deleteLog(pipelineLogId);
+    public void deleteLog(String logId) {
+        PipelineExecLog oneLog = findOneLog(logId);
+        if (oneLog != null){
+            pipelineCodeLogService.deleteCodeLog(oneLog);
+        }
+        pipelineExecLogDao.deleteLog(logId);
     }
 
     //更新
     @Override
     public void updateLog(PipelineExecLog pipelineExecLog) {
         PipelineExecLogEntity pipelineExecLogEntity = BeanMapper.map(pipelineExecLog, PipelineExecLogEntity.class);
+        pipelineCodeLogService.updateCodeLog(pipelineExecLog);
         pipelineExecLogDao.updateLog(pipelineExecLogEntity);
     }
 
     //查询单个
     @Override
-    public PipelineExecLog findOneLog(String pipelineLogId) {
-        PipelineExecLogEntity pipelineExecLogEntity = pipelineExecLogDao.findOne(pipelineLogId);
-        PipelineExecLog pipelineExecLog = BeanMapper.map(pipelineExecLogEntity, PipelineExecLog.class);
+    public PipelineExecLog findOneLog(String logId) {
+        PipelineExecLog pipelineExecLog = BeanMapper.map(pipelineExecLogDao.findOne(logId), PipelineExecLog.class);
         joinTemplate.joinQuery(pipelineExecLog);
         return pipelineExecLog;
     }
@@ -59,10 +89,10 @@ public class PipelineExecLogServiceImpl implements PipelineExecLogService {
     //查询所有
     @Override
     public List<PipelineExecLog> findAllLog() {
-        List<PipelineExecLogEntity> pipelineExecLogEntityList = pipelineExecLogDao.findAllLog();
-        List<PipelineExecLog> pipelineExecLogs = BeanMapper.mapList(pipelineExecLogEntityList, PipelineExecLog.class);
-        joinTemplate.joinQuery(pipelineExecLogEntityList);
-        return pipelineExecLogs;
+        List<PipelineExecLogEntity> allLog = pipelineExecLogDao.findAllLog();
+        List<PipelineExecLog> pipelineExecLogList = BeanMapper.mapList(allLog, PipelineExecLog.class);
+        joinTemplate.joinQuery(pipelineExecLogList);
+        return pipelineExecLogList;
     }
 
     @Override
